Tidy up intervals script
